+++ CHANGELOG.md
@@ -4,11 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [2.1.3](https://gitlab.com/AquaticNightmare/bundle_exporter/-/releases/2_1_3)
+## [2.2.0](https://gitlab.com/AquaticNightmare/bundle_exporter/-/releases/2_2_2)
 ### Added
 - Option for merge meshes to merge UVs by index instead of by name
 - New Pivot option "From Collection" it will get the pivot from the collection instance_offset parameter (Properties->Object->Collection->X,Y,Z)
 - Option to exclude collisions from the merge operation
+- New modifiers:
+    - Bake animations (Necessary for processing other armature modifiers)
+    - Delete bones (option to delete bones by name or if they don't use deform)
+
 ### Changed
 - Modifiers are now displayed in the order they are applied